requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.7
distro==1.8.0
Django==3.2.18
django-instagram==0.3.2
djangorestframework==3.14.0
filelock==3.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
jdcal==1.4.1
lxml==4.9.3
openai==1.3.2
openpyxl==2.0.5
Pillow==10.1.0
pipenv==2023.7.23
platformdirs==3.10.0
pydantic==2.5.1
pydantic_core==2.14.3
pytz==2022.7.1
requests==2.31.0
sniffio==1.3.0
sorl-thumbnail==12.10.0
soupsieve==2.5
sqlparse==0.4.3
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
